move helper class to utils
